fix: move suplly manager check to util
