+++ kit/dapp/src/app/[locale]/(private)/assets/[assettype]/[address]/(details)/_components/related/bonds.tsx
@@ -1,12 +1,12 @@
 import { BurnForm } from "@/app/[locale]/(private)/portfolio/my-assets/[assettype]/[address]/_components/burn-form/form";
 import { RelatedGrid } from "@/components/blocks/related-grid/related-grid";
 import { RelatedGridItem } from "@/components/blocks/related-grid/related-grid-item";
-import { ROLES } from "@/lib/config/roles";
 import { getAssetBalanceDetail } from "@/lib/queries/asset-balance/asset-balance-detail";
 import { getAssetDetail } from "@/lib/queries/asset-detail";
 import { getAssetUsersDetail } from "@/lib/queries/asset/asset-users-detail";
+import { isSupplyManager } from "@/lib/utils/has-role";
 import { getTranslations } from "next-intl/server";
-import { getAddress, type Address } from "viem";
+import type { Address } from "viem";
 import { MintForm } from "../../../_components/mint-form/form";
 
 interface BondsRelatedProps {
@@ -29,33 +29,10 @@ export async function BondsRelated({
   const isBlocked = userBalance?.blocked ?? false; // Still useful for general blocking
   const isPaused = "paused" in assetDetails && assetDetails.paused;
 
-  // Determine if user is supply manager using assetUsersDetails first
-  let userIsSupplyManager = false;
-
-  if (currentUserWallet && assetUsersDetails) {
-    try {
-      const normalizedUserAddress = getAddress(currentUserWallet);
-
-      const userRoleInfo = assetUsersDetails.roles.find((role) => {
-        try {
-          return getAddress(role.id) === normalizedUserAddress;
-        } catch {
-          return false;
-        }
-      });
-
-      const userRoles = userRoleInfo?.roles ?? [];
-
-      userIsSupplyManager = userRoles.includes(
-        ROLES.SUPPLY_MANAGEMENT_ROLE.contractRole
-      );
-    } catch (error) {
-      console.error(
-        "Error determining supply manager role from assetUsersDetails:",
-        error
-      );
-    }
-  }
+  const userIsSupplyManager = isSupplyManager(
+    currentUserWallet,
+    assetUsersDetails
+  );
 
   return (
     <RelatedGrid title={t("title")}>
