+++ kit/dapp/src/app/[locale]/(private)/assets/[assettype]/[address]/(details)/_components/related/deposits.tsx
@@ -1,13 +1,13 @@
 import { BurnForm } from "@/app/[locale]/(private)/portfolio/my-assets/[assettype]/[address]/_components/burn-form/form";
 import { RelatedGrid } from "@/components/blocks/related-grid/related-grid";
 import { RelatedGridItem } from "@/components/blocks/related-grid/related-grid-item";
-import { ROLES } from "@/lib/config/roles";
 import { getAssetBalanceDetail } from "@/lib/queries/asset-balance/asset-balance-detail";
 import { getAssetUsersDetail } from "@/lib/queries/asset/asset-users-detail";
 import type { getDepositDetail } from "@/lib/queries/deposit/deposit-detail";
+import { isSupplyManager } from "@/lib/utils/has-role";
 import { isBefore } from "date-fns";
 import { getTranslations } from "next-intl/server";
-import { getAddress, type Address } from "viem";
+import type { Address } from "viem";
 import { UpdateCollateralForm } from "../../../_components/manage-dropdown/update-collateral-form/form";
 import { MintForm } from "../../../_components/mint-form/form";
 
@@ -31,33 +31,10 @@ export async function DepositsRelated({
   const isBlocked = userBalance?.blocked ?? false;
   const isPaused = "paused" in assetDetails && assetDetails.paused;
 
-  // Determine if user is supply manager using assetUsersDetails first
-  let userIsSupplyManager = false;
-
-  if (currentUserWallet && assetUsersDetails) {
-    try {
-      const normalizedUserAddress = getAddress(currentUserWallet);
-
-      const userRoleInfo = assetUsersDetails.roles.find((role) => {
-        try {
-          return getAddress(role.id) === normalizedUserAddress;
-        } catch {
-          return false;
-        }
-      });
-
-      const userRoles = userRoleInfo?.roles ?? [];
-
-      userIsSupplyManager = userRoles.includes(
-        ROLES.SUPPLY_MANAGEMENT_ROLE.contractRole
-      );
-    } catch (error) {
-      console.error(
-        "Error determining supply manager role from assetUsersDetails:",
-        error
-      );
-    }
-  }
+  const userIsSupplyManager = isSupplyManager(
+    currentUserWallet,
+    assetUsersDetails
+  );
 
   const collateralIsExpired =
     "collateralProofValidity" in assetDetails &&
@@ -77,7 +54,9 @@ export async function DepositsRelated({
           address={address}
           assettype="deposit"
           asButton
-          disabled={isBlocked || isPaused || !userIsSupplyManager}
+          disabled={
+            isBlocked || isPaused || !userIsSupplyManager || collateralIsExpired
+          }
           decimals={deposit.decimals}
           symbol={deposit.symbol}
         />
