+++ kit/dapp/src/app/[locale]/(private)/assets/[assettype]/[address]/_components/mint-form/form.tsx
@@ -101,7 +101,7 @@ export function MintForm({
             <span tabIndex={0}>{formSheet}</span>
           </TooltipTrigger>
           <TooltipContent>
-            <p>Supply manager role required</p>
+            <p>{t("errors.supply-manager-required")}</p>
           </TooltipContent>
         </Tooltip>
       </TooltipProvider>
